add main components to be able to make slides
- background
- text
- blend background with black when text is shown
- toggle text layer on and off

Struggled a bit with the transparent color, which is always the 1st color in the
palette. For the sprite pixel art this is easy as you can control the index of
the 1st color yourself. Backgrounds are less controllable when using
photos. There's probably an easy trick to swap out this color in pngs with say
Aseprite or ImageMagick, but I couldn't figure it out so quick. Regardless you
still need to find out which color is closest to black as that's the color you
need for the background as that color is the color that doesn't blend with the
background. The easiest way seems to be to just check the end result palette in
the mGBA pallette viewer and then use Grit to swap out the desirable transparent
index and the 1st index color with -pT<index>. Seems a bit hacky, and this means
you need to have Makefile be specific about every single BG, but it'll have to
do for now.