fix messy LDFLAGS setup
For Grit we need to set LDFLAGS to point to libfreeimage. Previously we required
to set LDFLAGS from the cmdline and also set variables from the
CMakeLists.txt. To be a bit more hygenic, set LDFLAGS only for Grit
specifically in CMakeLists.txt and allow them to be overwritten from the
cmdline.