Rename lldb-vscode to lldb-dap in Debugging doc
It was renamed to better signal that it is generic,
not just for VS Code.

https://github.com/llvm/llvm-project/pull/69264

Also when I ported some of these instructions back
to LLVM, reviewers suggested I use the "attach" mode
instead.

https://github.com/llvm/llvm-project/tree/main/lldb/tools/lldb-dap#connect-to-a-debug-server-on-the-current-machine

So I've done that too.

For our purposes it's the same effect, but
technically we are attaching not launching.